+++ src/ValueObject/src/Integrations/HydratorStrategy/DateValueLaminasHydratorStrategy.php
@@ -36,10 +36,7 @@ final class DateValueLaminasHydratorStrategy implements StrategyInterface
 
     private function validateDateClass(string $dateClass): void
     {
-        $allowedClasses = [
-            DateTimeImmutableValue::class,
-            DateTimeValue::class,
-        ];
+        $allowedClasses = [DateTimeImmutableValue::class, DateTimeValue::class];
 
         foreach ($allowedClasses as $allowedClass) {
             if ($dateClass === $allowedClass || is_subclass_of($dateClass, $allowedClass)) {