+++ src/ValueObject/src/StringValue.php
@@ -6,26 +6,29 @@ namespace spaceonfire\ValueObject;
 
 abstract class StringValue extends BaseValueObject
 {
+    /**
+     * @inheritDoc
+     */
     protected function validate($value): bool
     {
         return is_scalar($value) || (is_object($value) && method_exists($value, '__toString'));
     }
 
+    /**
+     * @inheritDoc
+     * @return string
+     */
     protected function cast($value): string
     {
         return (string)$value;
     }
 
+    /**
+     * @inheritDoc
+     * @return string
+     */
     public function value(): string
     {
         return parent::value();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function jsonSerialize()
-    {
-        return (string)$this;
     }
 }