+++ src/ValueObject/src/UriValue.php
@@ -1,9 +1,5 @@
 <?php
 
-/**
- * @phpcs:disable PSR12.Classes.ClosingBrace.StatementAfter
- */
-
 declare(strict_types=1);
 
 namespace spaceonfire\ValueObject;
@@ -14,22 +10,30 @@ use Psr\Http\Message\UriInterface;
 
 class UriValue extends BaseValueObject
 {
+    /**
+     * @inheritDoc
+     */
     protected function validate($value): bool
     {
         return (is_string($value) && parse_url($value) !== false) || $value instanceof UriInterface;
     }
 
-    protected function cast($value)
+    /**
+     * @inheritDoc
+     * @return UriInterface
+     */
+    protected function cast($value): UriInterface
     {
         return !$value instanceof UriInterface ? new Uri($value) : $value;
     }
 
+    /**
+     * @inheritDoc
+     */
     protected function throwExceptionForInvalidValue(?string $value): void
     {
         if ($value !== null) {
-            throw new InvalidArgumentException(
-                sprintf('Expected a value to be a valid uri. Got "%s"', $value)
-            );
+            throw new InvalidArgumentException(sprintf('Expected a value to be a valid uri. Got "%s"', $value));
         }
 
         parent::throwExceptionForInvalidValue($value);
@@ -37,17 +41,10 @@ class UriValue extends BaseValueObject
 
     /**
      * @inheritDoc
+     * @return UriInterface
      */
     public function value(): UriInterface
     {
         return parent::value();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function jsonSerialize()
-    {
-        return (string)$this->value();
     }
 }