+++ src/ValueObject/tests/EmailValueTest.php
@@ -8,19 +8,19 @@ use PHPUnit\Framework\TestCase;
 
 class EmailValueTest extends TestCase
 {
-    public function testConstructor()
+    public function testConstructor(): void
     {
         $val = new EmailValue('test@test.test');
         $this->assertEquals('test@test.test', $val->value());
     }
 
-    public function testConstructFail()
+    public function testConstructFail(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         new EmailValue('just string');
     }
 
-    public function testConstructFailWithObject()
+    public function testConstructFailWithObject(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         new EmailValue(new \stdClass());