+++ src/ValueObject/tests/EnumValueTest.php
@@ -8,9 +8,19 @@ use PHPUnit\Framework\TestCase;
 
 class EnumValueTest extends TestCase
 {
+    /**
+     * @param mixed $val
+     * @return EnumValue
+     */
     private function factory($val): EnumValue
     {
-        return new class ($val) extends EnumValue {
+        /**
+         * @method static self one()
+         * @method static self twoWords()
+         * @method static self manyManyWords()
+         * @method static self oneString()
+         */
+        return new class($val) extends EnumValue {
             public const ONE = 1;
             public const TWO_WORDS = 2;
             public const MANY_MANY_WORDS = 3;
@@ -18,7 +28,7 @@ class EnumValueTest extends TestCase
         };
     }
 
-    public function testConstructor()
+    public function testConstructor(): void
     {
         $val = $this->factory(1);
         $this->assertEquals(1, $val->value());
@@ -26,19 +36,19 @@ class EnumValueTest extends TestCase
         $this->assertEquals('"1"', json_encode($val));
     }
 
-    public function testConstructorException()
+    public function testConstructorException(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->factory(4);
     }
 
-    public function testConstructFailWithObject()
+    public function testConstructFailWithObject(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->factory(new \stdClass());
     }
 
-    public function testEquals()
+    public function testEquals(): void
     {
         $a = $this->factory(1);
         $b = $this->factory(1);
@@ -51,7 +61,7 @@ class EnumValueTest extends TestCase
         $this->assertTrue($a->equals($b));
     }
 
-    public function testRandom()
+    public function testRandom(): void
     {
         $enum = $this->factory(1);
 
@@ -59,7 +69,7 @@ class EnumValueTest extends TestCase
         $this->assertContains($enum::random()->value(), $enum::values());
     }
 
-    public function testMagicCalls()
+    public function testMagicCalls(): void
     {
         $enum = $this->factory(1);