+++ src/ValueObject/tests/Integrations/HydratorStrategy/ValueObjectLaminasHydratorStrategyTest.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace spaceonfire\ValueObject\Integrations\HydratorStrategy;
+
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
+use spaceonfire\ValueObject\IpValue;
+
+class ValueObjectLaminasHydratorStrategyTest extends TestCase
+{
+    public function testConstruct(): void
+    {
+        new ValueObjectLaminasHydratorStrategy(IpValue::class);
+        self::assertTrue(true);
+    }
+
+    public function testConstructFail(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        new ValueObjectLaminasHydratorStrategy(\stdClass::class);
+    }
+
+    public function testExtract(): void
+    {
+        $strategy = new ValueObjectLaminasHydratorStrategy(IpValue::class);
+
+        $extracted = $strategy->extract(new IpValue('127.0.0.1'));
+        self::assertEquals('127.0.0.1', $extracted);
+    }
+
+    public function testHydrate(): void
+    {
+        $strategy = new ValueObjectLaminasHydratorStrategy(IpValue::class);
+
+        $hydrated = $strategy->hydrate('127.0.0.1', null);
+        self::assertEquals('127.0.0.1', $hydrated->value());
+    }
+}