+++ src/ValueObject/tests/StringValueTest.php
@@ -10,11 +10,11 @@ class StringValueTest extends TestCase
 {
     private function factory($val): StringValue
     {
-        return new class ($val) extends StringValue {
+        return new class($val) extends StringValue {
         };
     }
 
-    public function testConstruct()
+    public function testConstruct(): void
     {
         $val = $this->factory('Hello');
         $this->assertEquals('Hello', $val->value());
@@ -22,7 +22,7 @@ class StringValueTest extends TestCase
         $this->assertEquals('"Hello"', json_encode($val));
     }
 
-    public function testConstructFromNumber()
+    public function testConstructFromNumber(): void
     {
         $val = $this->factory(12345);
         $this->assertEquals(12345, $val->value());
@@ -30,7 +30,7 @@ class StringValueTest extends TestCase
         $this->assertEquals('"12345"', json_encode($val));
     }
 
-    public function testConstructFailWithObject()
+    public function testConstructFailWithObject(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->factory(new \stdClass());